pyproject.toml
[project]
name = "agent-monitor"
version = "0.1.0"
description = "Device runtime status and agent framework monitoring system"
requires-python = ">=3.11"
dependencies = [
    "psutil>=5.9",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
    "jinja2>=3.1",
    "websockets>=11.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
nvidia = ["pynvml>=11.5"]
alerting = ["plyer>=2.1"]

[project.scripts]
agent-monitor = "agent_monitor.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
